chore(jsx): rename file
